Move the "uuid" function to common/util.js and use the already-installed "uuid" module.
